requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
datasets==4.4.0
dill==0.4.0
fastapi==0.121.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
markupsafe==3.0.3
mlx==0.29.3
mlx-lm==0.28.3
mlx-metal==0.29.3
mlx-vlm==0.3.5
multidict==6.7.0
multiprocess==0.70.18
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pip==25.1.1
propcache==0.4.1
protobuf==6.33.0
pyarrow==22.0.0
pyautogui==0.9.54
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.1
starlette==0.49.3
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
xxhash==3.6.0
yarl==1.22.0
